+++ stonefish_slam/utils/visualization.py
@@ -8,6 +8,12 @@ from geometry_msgs.msg import Point
 from std_msgs.msg import ColorRGBA
 from visualization_msgs.msg import Marker, MarkerArray
 
+__all__ = [
+    'ros_colorline_trajectory',
+    'make_color_rgba',
+    'ros_constraints',
+]
+
 
 def ros_colorline_trajectory(traj):
     # ROS2: Create PointField objects properly